Add filter bar. This moves the search capability closer to the articles which makes a lot more sense. Cmd+Shift+B toggles the filter bar on/off, as does the 'f' shortcut key as it did before. Otherwise it works exactly as it did in previous builds plus the bonus of making the filter options more visible. The filter bar is on by default but is turned off when switching layouts.
The search field on the toolbar now works globally and populates a new Search
Results pseudo-folder with the results. The search string isn't currently
preserved across sessions.

This is functionally all complete but needs some polish and there's one bug
I've yet to fix relating to the condensed layout screwing up width when
switching from unified layout to condensed with the filter bar on.



git-svn-id: https://vienna-rss.svn.sourceforge.net/svnroot/vienna-rss/trunk@937 09869ea3-740d-0410-b225-97ef48a472b4