Improved toolbar code. Pressed state images are now handled distinctively. Small button images can now be specified separately for cases where Cocoa doesn't do that good a job of resizing and rescaling the larger image.
Still need improvements to the actual buttons as they look a bit too dark.
And small size toolbar looks ugly to the point where maybe we should
consider disabling small size entirely.

Also fix a crashing bug with downloading non-enclosure files and fix the
behaviour of the search bar with an empty string.



git-svn-id: https://vienna-rss.svn.sourceforge.net/svnroot/vienna-rss/trunk@954 09869ea3-740d-0410-b225-97ef48a472b4