Cache the background images. This fixes the black toolbar bug for now but it isn't what I'd consider to be a real fix as I haven't been able to identify where the image is being corrupted.
The problem was appearing with setting bgColor from the top_middle.png. This
is normally a light grey colour but after a few iterations, it becomes black.
I think we're hitting some resource issue.

Also, with this fix, I'm still seeing a painting issue. After a little while,
the top middle row ended up being painted with a while sequence of spinner
controls along the top.



git-svn-id: https://vienna-rss.svn.sourceforge.net/svnroot/vienna-rss/trunk@962 09869ea3-740d-0410-b225-97ef48a472b4