This looks to be a better way to do the polished metal background.
Replace PolishedWindow with a background brush based on mainBackground.tiff.
This is set for both the main Vienna window and the download window. I've kept
the old PolishedWindow code in for now just in case we need to revert back to
it for some reason but if this works then it should be deleted and the two
acknowledgement files updated.

So far I've seen no corruption in either window in release build.



git-svn-id: https://vienna-rss.svn.sourceforge.net/svnroot/vienna-rss/trunk@979 09869ea3-740d-0410-b225-97ef48a472b4